Move User form types to Core
